Error handling Added

+++ backend/routes/posts.js
@@ -34,7 +34,8 @@ router.post("",checkAuth,multer({storage : storage}).single("image"),(req,res,ne
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator : req.userData.userId
   });
   post.save().then(createdPost =>{
     res.status(201).json({
@@ -43,6 +44,11 @@ router.post("",checkAuth,multer({storage : storage}).single("image"),(req,res,ne
         ...createdPost,
         id: createdPost._id
       }
+    });
+  })
+  .catch(error => {
+    res.status(500).json({
+      message : "Post Creation Failed!"
     });
   });
 });
@@ -60,11 +66,23 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator : req.userData.userId
     });
     console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id , creator: req.userData.userId}, post).then(result => {
+      console.log(result);
+      if(result.matchedCount > 0){
+        res.status(200).json({ message: "Update successful!" });
+      }
+      else {
+        res.status(401).json({ message: "Not Authorized" });
+      }
+
+    }).catch(error => {
+      res.status(500).json({
+        message : "Couldn't Update Post!"
+      });
     });
   }
 );
@@ -88,6 +106,10 @@ router.get("", (req, res, next) => {
       posts: fetchedPost,
       maxPosts : count
     });
+  }).catch(error => {
+    res.status(500).json({
+      message : "Fetching Posts Failed!"
+    });
   });
 });
 
@@ -98,13 +120,26 @@ router.get("/:id", (req, res, next) => {
     } else {
       res.status(404).json({ message: "Post not found!" });
     }
+  }).catch(error => {
+    res.status(500).json({
+      message : "Fetching Post Failed!"
+    });
   });
 });
 
 router.delete("/:id",checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
+  Post.deleteOne({ _id: req.params.id , creator : req.userData.userId}).then(result => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+    if(result.deletedCount > 0){
+      res.status(200).json({ message: "Deletion successful!" });
+    }
+    else {
+      res.status(401).json({ message: "Not Authorized" });
+    }
+  }).catch(error => {
+    res.status(500).json({
+      message : "Posts Deletion Failed!"
+    });
   });
 });
